// frontend/felcn-base-frontend/src/app/(fase_2)/apd/registro_caso/types/registro-caso-apd.types.ts
export interface CatalogOption<T = unknown> {
  value: string
  label: string
  original: T
}

export interface AsignacionLgiDetalle {
  casosId: string
  dptoavId: string
  idGrupo: number | null
  uniAbrev: string | null
  disId: string
  descripcionGrupo?: string | null
  puesto?: string | null
  unidad?: string | null
  regional?: string | null
  idUnidad?: number | null
  nombreCaso: string
  nroCasoGiaef: string | null
  nroCaso: string
  nroCasoFis: string | null
  cudifp: string | null
  perddom: boolean | null
  nroCasoPerdom: string | null
  idEtapa: string | number | null
  remiteFiscal: string
  fechaRecepcionFiscalia: string | null
  conformeA: string
  inicioCaso: string | null
  codigoServicio: string | null
  fechaInicio: string | null
  idEstado: number | null
  diasOtorgados: number | null
  controlJurisdiccional?: string | null
  estado: string | null
  fechaHoraIng: string | null
  usuario: string | null
  usuarioActualizacion: string | null
  fechaActualizacion: string | null
}